requirements.txt
cryptography>=41.0.0
paho-mqtt<2.0.0